rm'd kernel_rpc, created IPC folder, created ring-chan dir:
IPC:
IPC needs to be its own stand-alone set of functions which just do IPC. Before in kernel_rpc as well as in the kernel/betamodule set of misc drivers the tests were incorporated into the IPC code itself. It was then driven from a userland program. In reality this isn't anywhere close to what we need to accomplish. Because of this we abstract all the true IPC code into its own .o file which can be linked in and used in future modules.

ring-chan:
Due to the re-write I've started re-writing the ring-channel code as well as the actual implementation of the ring-channels. There were way too many variables we were not using as well as levels of pointer chasing to get sizes etc which is unecessary for us, and slow.

TODO:
Finish ring-chan cleanup
rewrite make files
setup tests dir -- essentially an RPC mechanism
research intel HLE to see if it holds the key to our ~200cycle IPC on a single line

Signed-off-by: Scotty Bauer <sbauer@eng.utah.edu>