Added new function to IPC API: attach_thread_to_channel.
Split create_channel into two functions:
create_channel simply allocates space for the channel and
the ring buffer.

attach_thread_to_channel:
This function creates a kthread and pins it to a specific core.
It takes a funciton pointer to the function you want to execute.

**NOTE**
The timing code is still using the API, It will be fixed in future
commits

Signed-off-by: Scott Bauer <sbauer@eng.utah.edu>