lcd: libfipc builds.
I needed to make the struct fipc_ring_buf a bit bigger -- 2 cache
lines instead of 1. This is because the kernel's struct mutex
can be a little big, depending on the config. If we really want
to make it smaller, we should require the kernel config to be
set a certain way, perhaps.